requirements.txt
absl-py==1.4.0
aiofiles==22.1.0
aiohttp==3.8.4
aiosignal==1.3.1
aiosqlite==0.18.0
alembic==1.10.2
altair==4.2.2
annotated-types==0.5.0
anyio==3.6.2
appdirs==1.4.4
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
asttokens==2.2.1
astunparse==1.6.3
async-generator==1.10
async-timeout==4.0.2
attrs==23.1.0
Babel==2.12.1
backcall==0.2.0
batchspawner==1.2.0
beautifulsoup4==4.12.0
bleach==6.0.0
blinker==1.6.2
blis==0.7.11
blosc2==2.0.0
bokeh==3.1.0
boto3==1.26.115
botocore==1.29.115
cachetools==5.3.0
catalogue==2.0.10
certifi==2022.12.7
certipy==0.1.3
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
cloudpickle==2.2.1
cmake==3.26.0
comm==0.1.3
confection==0.1.3
contourpy==1.0.7
cryptography==39.0.2
cycler==0.11.0
cymem==2.0.8
Cython==0.29.33
dask==2023.3.1
dask-jobqueue==0.8.1
dask-labextension==6.1.0
datasets==2.11.0
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
dill==0.3.6
distlib==0.3.6
distributed==2023.3.1
docker-pycreds==0.4.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.12.0
flatbuffers==23.3.3
fonttools==4.39.2
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.4.0
gast==0.4.0
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.17.3
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
greenlet==2.0.2
grpcio==1.54.0
h5py==3.8.0
HeapDict==1.0.1
huggingface-hub==0.13.4
idna==3.4
imageio==2.33.1
importlib-metadata==6.5.0
importlib-resources==5.12.0
ipykernel==6.22.0
ipympl==0.9.3
ipython==8.11.0
ipython-genutils==0.2.0
ipywidgets==8.0.5
isoduration==20.11.0
jax==0.4.6
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
json5==0.9.11
jsonpointer==2.3
jsonschema==4.17.3
jupyter-events==0.6.3
jupyter-server-proxy==3.2.2
jupyter-telemetry==0.1.0
jupyter-ydoc==0.2.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_fileid==0.8.0
jupyter_server_terminals==0.4.4
jupyter_server_ydoc==0.8.0
jupyterhub==3.1.1
jupyterlab==3.6.2
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.6
jupyterlab_server==2.21.0
jupyterlmod==3.0.1
keras==2.12.0
kiwisolver==1.4.4
langcodes==3.3.0
lazy_loader==0.3
libclang==15.0.6.1
lime==0.2.0.1
lit==16.0.0
locket==1.0.0
Mako==1.2.4
Markdown==3.4.3
markdown-it-py==2.2.0
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mdurl==0.1.2
meson==1.3.2
mistune==2.0.5
mpmath==1.3.0
msgpack==1.0.5
multidict==6.0.4
multiprocess==0.70.14
murmurhash==1.0.10
mysql==0.0.3
mysqlclient==2.1.1
nbclassic==0.5.3
nbclient==0.7.2
nbconvert==7.2.10
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numexpr==2.8.4
numpy==1.24.2
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-cupti-cu11==11.7.101
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
nvidia-cufft-cu11==10.9.0.58
nvidia-curand-cu11==10.2.10.91
nvidia-cusolver-cu11==11.4.0.1
nvidia-cusparse-cu11==11.7.4.91
nvidia-nccl-cu11==2.14.3
nvidia-nvtx-cu11==11.7.91
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.1
pamela==1.0.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
partd==1.3.0
pathtools==0.1.2
pathy==0.10.2
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.1.1
plotly==5.13.1
preshed==3.0.9
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.5
psycopg2-binary==2.9.5
ptyprocess==0.7.0
pure-eval==0.2.2
py-cpuinfo==9.0.0
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydantic==2.4.0
pydantic_core==2.10.0
pydeck==0.8.1b0
Pygments==2.15.1
Pympler==1.0.1
pyOpenSSL==23.0.0
pyparsing==3.0.9
pyro-api==0.1.2
pyro-ppl==1.8.4
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
python-version==0.0.2
pytorch-pretrained-bert==0.6.2
pytz==2023.3
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
pyzmq==25.0.2
randtest @ git+https://github.com/estripling/randtest@1cfa746520e426b664aae2518f74462f810d6d84
regex==2023.3.23
requests==2.28.2
requests-oauthlib==1.3.1
responses==0.18.0
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rich==13.3.4
rsa==4.9
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
s3transfer==0.6.0
scikit-image==0.22.0
scikit-learn==1.2.2
scipy==1.10.1
seaborn==0.12.2
semver==3.0.0
Send2Trash==1.8.0
sentencepiece==0.1.98
sentry-sdk==1.20.0
seqeval==1.2.2
setproctitle==1.3.2
simpervisor==0.4
simpletransformers==0.63.9
six==1.16.0
smart-open==6.4.0
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4
spacy==3.6.1
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.7
srsly==2.4.8
stack-data==0.6.2
streamlit==1.12.0
sympy==1.11.1
tables==3.8.0
tblib==1.7.0
tenacity==8.2.2
tensorboard==2.12.2
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
tensorflow==2.12.0
tensorflow-estimator==2.12.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.2.0
terminado==0.17.1
thinc==8.1.12
threadpoolctl==3.1.0
tifffile==2023.12.9
tinycss2==1.2.1
tokenizers==0.13.3
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch==2.0.0
tornado==6.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
triton==2.0.0
typer==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==4.3
uri-template==1.2.0
urllib3==1.26.15
validators==0.20.0
virtualenv==20.21.0
wandb==0.14.2
wasabi==1.1.2
watchdog==3.0.0
wcwidth==0.2.6
webcolors==1.12
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.6
wordcloud==1.9.3
wrapt==1.14.1
xlrd==2.0.1
XlsxWriter==3.0.9
xlwt==1.3.0
xxhash==3.2.0
xyzservices==2023.2.0
y-py==0.5.9
yarl==1.8.2
ypy-websocket==0.8.2
zict==2.2.0
zipp==3.15.0
